requirements.txt
pycryptodome
pillow